+++ frontend/app/src/entities/events/ui/event.tsx
@@ -2,16 +2,10 @@ import { EventNodeInterface, NodeMutatedEvent } from "@/shared/api/graphql/gener
 import { DateDisplay } from "@/shared/components/display/date-display";
 
 import { DisplayLabel } from "@/entities/nodes/object/ui/display-label";
-import { schemaKindLabelState } from "@/entities/schema/stores/schemaKindLabel.atom";
 import { PropertyRow } from "@/entities/schema/ui/styled";
-import { constructPath } from "@/shared/api/rest/fetch";
 import { CopyToClipboard } from "@/shared/components/buttons/copy-to-clipboard";
-import { Link } from "@/shared/components/ui/link";
 import { Popover, PopoverContent, PopoverTrigger } from "@/shared/components/ui/popover";
 import { TimelineBorder } from "@/shared/components/ui/timeline-border";
-import { Icon } from "@iconify-icon/react";
-import { useAtomValue } from "jotai";
-import { ReactElement } from "react";
 import {
   BRANCH_CREATED_EVENT,
   BRANCH_DELETED_EVENT,
@@ -19,6 +13,8 @@ import {
   BRANCH_REBASEDED_EVENT,
   NODE_MUTATED_EVENT,
 } from "../utils/constants";
+import { BranchEvent } from "./branch-event";
+import { EventAttributes, NodeEvent } from "./node-event";
 
 export type BranchEventType = EventNodeInterface & {
   __typename:
@@ -32,52 +28,6 @@ export type NodeEventType = NodeMutatedEvent & {
 };
 
 export type EventType = BranchEventType | NodeEventType;
-
-export const NODE_EVENTS_MAPPING: Record<string, string> = {
-  "infrahub.node.created": "created",
-  "infrahub.node.updated": "updated",
-  "infrahub.node.deleted": "deleted",
-};
-
-export const BRANCH_EVENTS_MAPPING: Record<string, (param: string) => ReactElement> = {
-  "infrahub.branch.created": (branch) => (
-    <div>
-      Branch <span className="text-black font-semibold">{branch}</span> created
-    </div>
-  ),
-  "infrahub.branch.rebased": (branch) => (
-    <div>
-      Branch <span className="text-black font-semibold">{branch}</span> rebased
-    </div>
-  ),
-  "infrahub.branch.deleted": (branch) => (
-    <div>
-      Branch <span className="text-black font-semibold">{branch}</span> deleted
-    </div>
-  ),
-};
-
-const EventAttributes = ({ attributes }: Pick<NodeMutatedEvent, "attributes">) => {
-  return (
-    <div className="pl-8 text-sm">
-      {attributes.map(({ action, name, value, value_previous }) => {
-        return (
-          <div className="grid grid-cols-2 gap-2" key={`${action}_${name}`}>
-            <div className="font-normal">{name}</div>
-
-            <div className="flex items-center gap-4">
-              <div className="text-gray-400">{value_previous ?? "Ø"}</div>
-
-              <Icon icon={"mdi:chevron-right"} className="text-custom-blue-500" />
-
-              <div>{value ?? "Ø"}</div>
-            </div>
-          </div>
-        );
-      })}
-    </div>
-  );
-};
 
 const EventDetails = ({ id, event, occurred_at, account_id, ...props }: EventType) => {
   return (
@@ -97,58 +47,6 @@ const EventDetails = ({ id, event, occurred_at, account_id, ...props }: EventTyp
         <PropertyRow title="Changes" value={<EventAttributes attributes={props.attributes} />} />
       )}
     </div>
-  );
-};
-
-const NodeEvent = (props: NodeMutatedEvent) => {
-  const { event, account_id } = props;
-  const schemaLabels = useAtomValue(schemaKindLabelState);
-
-  return (
-    <>
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-2 text-sm">
-          {account_id && (
-            <div className="font-semibold">
-              <DisplayLabel id={account_id} />
-            </div>
-          )}
-
-          <div className="text-gray-500">{NODE_EVENTS_MAPPING[event] ?? "-"}</div>
-
-          <div className="font-semibold">{schemaLabels[props.payload.data.node_kind] ?? "-"}</div>
-
-          <Link
-            to={constructPath(
-              `/objects/${props.payload.data.node_kind}/${props.payload.data.node_id}`
-            )}
-          >
-            <DisplayLabel id={props.payload.data.node_id} />
-          </Link>
-        </div>
-      </div>
-
-      <EventAttributes attributes={props.attributes} />
-    </>
-  );
-};
-
-const BranchEvent = (props: EventNodeInterface) => {
-  const { event, occurred_at, branch } = props;
-
-  return (
-    <>
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-2 text-sm">
-          <div className="text-gray-500">
-            {branch && BRANCH_EVENTS_MAPPING[event] && BRANCH_EVENTS_MAPPING[event](branch)}
-          </div>
-        </div>
-        <div className="text-xs font-medium text-gray-500 dark:text-neutral-400">
-          <DateDisplay date={occurred_at} />
-        </div>
-      </div>
-    </>
   );
 };
