+++ frontend/app/src/entities/events/ui/node-event.tsx
@@ -1,0 +1,68 @@
+import { DisplayLabel } from "@/entities/nodes/object/ui/display-label";
+import { schemaKindLabelState } from "@/entities/schema/stores/schemaKindLabel.atom";
+import { NodeMutatedEvent } from "@/shared/api/graphql/generated/graphql";
+import { constructPath } from "@/shared/api/rest/fetch";
+import { Link } from "@/shared/components/ui/link";
+import { Icon } from "@iconify-icon/react";
+import { useAtomValue } from "jotai";
+
+export const NODE_EVENTS_MAPPING: Record<string, string> = {
+  "infrahub.node.created": "created",
+  "infrahub.node.updated": "updated",
+  "infrahub.node.deleted": "deleted",
+};
+
+export const EventAttributes = ({ attributes }: Pick<NodeMutatedEvent, "attributes">) => {
+  return (
+    <div className="pl-8 text-sm">
+      {attributes.map(({ action, name, value, value_previous }) => {
+        return (
+          <div className="grid grid-cols-2 gap-2" key={`${action}_${name}`}>
+            <div className="font-normal">{name}</div>
+
+            <div className="flex items-center gap-4">
+              <div className="text-gray-400">{value_previous ?? "Ø"}</div>
+
+              <Icon icon={"mdi:chevron-right"} className="text-custom-blue-500" />
+
+              <div>{value ?? "Ø"}</div>
+            </div>
+          </div>
+        );
+      })}
+    </div>
+  );
+};
+
+export const NodeEvent = (props: NodeMutatedEvent) => {
+  const { event, account_id } = props;
+  const schemaLabels = useAtomValue(schemaKindLabelState);
+
+  return (
+    <>
+      <div className="flex items-center justify-between">
+        <div className="flex items-center gap-2 text-sm">
+          {account_id && (
+            <div className="font-semibold">
+              <DisplayLabel id={account_id} />
+            </div>
+          )}
+
+          <div className="text-gray-500">{NODE_EVENTS_MAPPING[event] ?? "-"}</div>
+
+          <div className="font-semibold">{schemaLabels[props.payload.data.node_kind] ?? "-"}</div>
+
+          <Link
+            to={constructPath(
+              `/objects/${props.payload.data.node_kind}/${props.payload.data.node_id}`
+            )}
+          >
+            <DisplayLabel id={props.payload.data.node_id} />
+          </Link>
+        </div>
+      </div>
+
+      <EventAttributes attributes={props.attributes} />
+    </>
+  );
+};
